+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add extra event context params and and tags promotion ([#183](https://github.com/getsentry/sentry-unreal/pull/183))
 - Add automatic crash capturing for Mac ([#190](https://github.com/getsentry/sentry-unreal/pull/190))
 - Add environment property to plugin settings ([#204](https://github.com/getsentry/sentry-unreal/pull/204))
+- Add native sources upload toggle to plugin settings ([#217](https://github.com/getsentry/sentry-unreal/pull/217))
 
 ### Fixes